+++ minchin/releaser/__init__.py
@@ -1,11 +1,6 @@
 # -*- coding: UTF-8 -*-
 
-__title__ = "minchin.releaser"
-__version__ = "0.3.1"
-__description__ = "Minchin.Releaser is a collection of tools designed to make releasing Python packages easier."
-__author__ = "William Minchin"
-__email__ = "w_minchin@hotmail.com"
-__url__ = "https://github.com/MinchinWeb/minchin.releaser"
-__license__ = "MIT License"
-
+from .constants import (__author__, __description__, __email__, __license__,
+                        __title__, __url__, __version__)
 from .make_release import make_release
+from .vendorize import vendorize